[tint][wgsl] Replace size_t with uint32_t for IDs, Source.

+++ src/tint/lang/wgsl/ast/node_id.h
@@ -46,7 +46,7 @@ struct NodeID {
     bool operator<(NodeID other) const { return value < other.value; }
 
     /// The numerical value for the node identifier
-    size_t value = 0;
+    uint32_t value = 0;
 };
 
 }  // namespace tint::ast
